backend/requirements.txt
fastapi>=0.124.0
uvicorn>=0.35.0
websockets>=15.0.1
google-genai>=1.0.0
python-dotenv>=1.1.0
